pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "federated-inference"
version = "0.1.0"
description = "Run large LLMs across multiple heterogeneous edge devices via pipeline parallelism"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "grpcio>=1.62",
    "protobuf>=4.25",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "click>=8.1",
]

[project.optional-dependencies]
coordinator = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
]
client = [
    "openai>=1.0",
]
all = [
    "federated-inference[coordinator,client]",
]
dev = [
    "grpcio-tools>=1.62",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx>=0.27",
]

[project.scripts]
federated-coordinator = "federated_inference.cli.coordinator_cli:main"
federated-worker = "federated_inference.cli.worker_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/federated_inference"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build]
sources = ["src"]
